# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Brain Tumor Detection System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, doctor, student]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /scans:
    get:
      operationId: getScans
      tags: [scans]
      summary: Get scans
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, analyzed, reviewed]
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScansListResponse"

    post:
      operationId: uploadScan
      tags: [scans]
      summary: Upload MRI scan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                patientId:
                  type: integer
                notes:
                  type: string
      responses:
        "201":
          description: Scan uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}:
    get:
      operationId: getScanById
      tags: [scans]
      summary: Get scan by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanWithAnalysis"
    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete scan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /scans/{id}/notes:
    post:
      operationId: addScanNote
      tags: [scans]
      summary: Add doctor note to scan
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddNoteRequest"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanNote"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, doctor, student]
        organization:
          type: string
        avatar:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - role
        - isActive
        - createdAt
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        organization:
          type: string
        role:
          type: string
          enum: [admin, doctor, student]
        isActive:
          type: boolean
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    Scan:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        status:
          type: string
          enum: [pending, analyzed, reviewed]
        patientId:
          type: integer
        patientName:
          type: string
        uploadedById:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - patientId
        - uploadedById
        - createdAt
        - updatedAt
    ScanNote:
      type: object
      properties:
        id:
          type: integer
        scanId:
          type: integer
        doctorId:
          type: integer
        doctorName:
          type: string
        note:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - doctorId
        - note
        - createdAt
    AddNoteRequest:
      type: object
      properties:
        note:
          type: string
      required:
        - note
    ScanWithAnalysis:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        status:
          type: string
          enum: [pending, analyzed, reviewed]
        patientId:
          type: integer
        patientName:
          type: string
        uploadedById:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        analysis:
          $ref: "#/components/schemas/AnalysisResult"
        doctorNotes:
          type: array
          items:
            $ref: "#/components/schemas/ScanNote"
      required:
        - id
        - status
        - patientId
        - uploadedById
        - createdAt
        - updatedAt
        - doctorNotes
    ScansListResponse:
      type: object
      properties:
        scans:
          type: array
          items:
            $ref: "#/components/schemas/Scan"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - scans
        - total
        - page
        - limit
    AnalysisResult:
      type: object
      properties:
        id:
          type: integer
        scanId:
          type: integer
        tumorType:
          type: string
          enum: [glioma, meningioma, pituitary, no_tumor]
        confidence:
          type: number
          format: float
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        summary:
          type: string
        keyFindings:
          type: array
          items:
            type: string
        analyzedAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - tumorType
        - confidence
        - riskLevel
        - analyzedAt
...