# lib/api-zod/src/generated/types/scanWithAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brain Tumor Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";
import type { ScanNote } from "./scanNote";
import type { ScanWithAnalysisStatus } from "./scanWithAnalysisStatus";

export interface ScanWithAnalysis {
  id: number;
  imageUrl?: string;
  status: ScanWithAnalysisStatus;
  patientId: number;
  patientName?: string;
  uploadedById: number;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
  analysis?: AnalysisResult;
  doctorNotes: ScanNote[];
}